feat: add new sync subcommand to playout generator
