+++ crates/ersatztv-playout-generator/src/main.rs
@@ -1,18 +1,11 @@
 mod error;
+mod generate;
 
-use std::collections::HashMap;
 use std::path::{Path, PathBuf};
 
-use clap::Parser;
-use ersatztv_playout::playout::{
-    DATE_FORMAT, Playout, PlayoutItem, PlayoutItemSource, PlayoutItemTracks, TrackSelection,
-};
-use ffpipeline::probe::{ProbeResult, ProbeResultStream};
-use rand::RngExt;
-use rand::seq::SliceRandom;
-use time::OffsetDateTime;
+use clap::{Parser, Subcommand};
+use ffpipeline::probe::ProbeResult;
 use walkdir::DirEntry;
-use walkdir::WalkDir;
 
 use crate::error::PlayoutGeneratorError;
 
@@ -24,12 +17,23 @@ static VIDEO_EXTENSIONS: &[&str] = &[
 static IMAGE_EXTENSIONS: &[&str] = &["png"];
 
 #[derive(Parser, Debug)]
-#[command(version, about, long_about = None)]
+#[command(version, about, long_about = None, subcommand_negates_reqs = true)]
 struct Args {
-    #[arg(short, long)]
-    content_folder: PathBuf,
-    #[arg(short, long)]
-    output_folder: PathBuf,
+    #[arg(short, long, required = true)]
+    content_folder: Option<PathBuf>,
+    #[arg(short, long, required = true)]
+    output_folder: Option<PathBuf>,
+
+    #[command(subcommand)]
+    command: Option<Commands>,
+}
+
+#[derive(Subcommand, Debug)]
+enum Commands {
+    Sync {
+        #[arg(short, long, required = true)]
+        database: PathBuf,
+    },
 }
 
 #[tokio::main]
@@ -45,125 +49,17 @@ pub async fn main() {
 async fn run() -> Result<(), PlayoutGeneratorError> {
     let args = Args::parse();
 
-    // find all available video files
-    let mut video_paths: HashMap<PathBuf, ProbeResult> = HashMap::new();
-    for path_and_probe in WalkDir::new(args.content_folder)
-        .into_iter()
-        .filter_map(|e| e.ok())
-        .filter(|e| e.file_type().is_file())
-        .filter(|e| is_video_extension(e) || is_image_extension(e))
-        .filter_map(to_probe_result)
-    {
-        log::debug!("path: {path_and_probe:?}");
-        video_paths.insert(path_and_probe.path, path_and_probe.probe);
-    }
-
-    if video_paths.is_empty() {
-        return Err(PlayoutGeneratorError::NoSourceContent);
-    }
-
-    // generate output file name
-    let start = OffsetDateTime::now_local()?.truncate_to_day();
-    let formatted_start = start.format(&DATE_FORMAT)?;
-    let finish = start + time::Duration::days(1) - time::Duration::seconds(1);
-
-    let video_list: Vec<(PathBuf, ProbeResult)> = video_paths.into_iter().collect();
-
-    // fill output file with content
-    let mut playout_items: Vec<PlayoutItem> = Vec::new();
-    let mut current_time = start;
-    let mut rng = rand::rng();
-    let mut shuffled = video_list.clone();
-    shuffled.shuffle(&mut rng);
-    let mut cursor = 0;
-
-    while current_time < finish {
-        if cursor >= shuffled.len() {
-            let last = shuffled.last().cloned();
-            shuffled.shuffle(&mut rng);
-            if shuffled.len() > 1
-                && let Some(ref last) = last
-                && shuffled[0].0 == last.0
-            {
-                let swap_index = rng.random_range(1..shuffled.len());
-                shuffled.swap(0, swap_index);
-            }
-            cursor = 0;
+    match args.command {
+        Some(Commands::Sync { database }) => {
+            log::info!("synchronizing from database at {:?}", database);
+            Ok(())
         }
-
-        let (path, probe_result) = &shuffled[cursor];
-        cursor += 1;
-
-        let is_image = probe_result.streams.len() == 1
-            && matches!(probe_result.streams.first(), Some(ProbeResultStream::Video(video_stream)) if IMAGE_EXTENSIONS.contains(&video_stream.codec.as_str()));
-
-        // use 10-sec duration for images
-        let image_duration = std::time::Duration::from_secs(10);
-        let duration = match probe_result.duration {
-            Some(d) => Some(d),
-            None => {
-                if is_image {
-                    Some(image_duration)
-                } else {
-                    None
-                }
-            }
-        };
-
-        if let Some(scheduled_duration) = duration
-            && let Ok(mut playout_item) = PlayoutItem::new(
-                uuid::Uuid::new_v4().to_string(),
-                current_time,
-                current_time + scheduled_duration,
-                None,
-                None,
-                path,
-            )
-        {
-            // use separate tracks with lavfi audio for images
-            if is_image {
-                playout_item.tracks = Some(PlayoutItemTracks {
-                    audio: Some(TrackSelection::Source {
-                        source: PlayoutItemSource::Lavfi {
-                            params: format!(
-                                "sine=frequency=1000:sample_rate=48000:d={}",
-                                image_duration.as_secs()
-                            ),
-                        },
-                        // source: PlayoutItemSource::Local {
-                        //     path: String::from("~/Music/silence.mp3"),
-                        //     in_point_ms: None,
-                        //     out_point_ms: Some(image_duration.as_millis() as u64),
-                        // },
-                    }),
-                    video: playout_item
-                        .source
-                        .as_ref()
-                        .map(|s| TrackSelection::Source { source: s.clone() }),
-                });
-
-                playout_item.source = None;
-            }
-
-            playout_items.push(playout_item);
-            current_time += scheduled_duration;
+        None => {
+            let content_folder = args.content_folder.unwrap();
+            let output_folder = args.output_folder.unwrap();
+            generate::generate_playout(&content_folder, &output_folder).await
         }
     }
-
-    let formatted_finish = current_time.format(&DATE_FORMAT)?;
-    let output_file = format!("{formatted_start}_{formatted_finish}.json");
-    log::debug!("output_file: {output_file}");
-
-    if !args.output_folder.exists() {
-        tokio::fs::create_dir_all(&args.output_folder).await?;
-    }
-
-    let output_path = args.output_folder.join(&output_file);
-    let playout = Playout::new(playout_items);
-    let output_string = serde_json::to_string_pretty(&playout)?;
-    tokio::fs::write(&output_path, output_string).await?;
-
-    Ok(())
 }
 
 fn is_video_extension(dir_entry: &DirEntry) -> bool {
